Move module util.graph: * util.graph -> geoanalytics.graph * move class CoordinateSpanningTree to geoanalytics.graph * update requirements.txt * update setup.py * add additional dependencies to auto_mock_imports in docs/conf.py
