+++ requirements.txt
@@ -4,4 +4,3 @@ numpy~=1.18.0
 scikit-learn~=0.22.0
 seaborn~=0.10.0
 typing-extensions==3.7.4.2
-networkx==2.4
