+++ setup.py
@@ -3,7 +3,7 @@ from setuptools import find_packages, setup
 tf_requirements = ['tensorflow==1.15.0']
 torch_requirements = ['torch==1.4.0', 'torchtext==0.5.0']
 lightgbm_requirements = ['lightgbm==2.3.0']
-geoanalytics_requirements = ['Shapely~=1.7.0', 'geopandas==0.7.0']
+geoanalytics_requirements = ['networkx==2.4', 'Shapely~=1.7.0', 'geopandas==0.7.0']
 
 setup(
     name='sensai',
